Enabling (and fixing) 64-bit builds.

+++ framework/Source/GPUImageFilterGroup.m
@@ -117,7 +117,7 @@
     }
 }
 
-- (NSInteger)nextAvailableTextureIndex;
+- (GLint)nextAvailableTextureIndex;
 {
 //    if ([_initialFilters count] > 0)
 //    {
